+++ lang/en/mumie.php
@@ -76,13 +76,11 @@ $string['gradetopassmustbeset'] = '<i>Grade to pass</i> cannot be zero as this a
 $string['mumie_form_due_date'] = 'Due date';
 $string['mumie_form_due_date_help'] = 'If enabled, grades that were earned after the selected date will not be synchronized with Moodle';
 $string['mumie_form_due_date_must_be_future'] = 'You must select a date in the future!';
-$string['mumie_form_grade_pool'] = 'Grade pool';
+$string['mumie_form_grade_pool'] = 'Share grades with other courses';
 $string['mumie_form_grade_pool_help'] = 'Choose whether to share a grade pool with other MOODLE courses or to use a private one.
 <br>If sharing is enabled, points that were earned for MUMIE problems in other courses will be automatically synchronized with this course\'s gradebook.
-<br>If private pool is selected, this course will neither be able to import nor to export grades.';
+<br>If not, this course will neither be able to import nor to export grades.';
 $string['mumie_form_grade_pool_warning'] = '<b style="color:red">Warning:</b><br> This decision is <b>final</b> and affects all other MUMIE Tasks in this course';
-$string['mumie_form_grade_pool_shared'] = 'Shared';
-$string['mumie_form_grade_pool_private'] = 'Private';
 
 // Used in mumieserver form.
 $string['mumie_form_required'] = 'required';