+++ changelog.md
@@ -12,6 +12,7 @@ All important changes to this plugin will be documented in this file.
 - When creating a MUMIE Task, MUMIE problems can be filtered by keywords.
 - A due date can now be set for MUMIE Tasks
 - Automatically add MUMIE Tasks by dragging and dropping MUMIE Problems from MUMIE courses into a MOODLE course.
+- Sharing grades for the same MUMIE problems with other MOODLE courses can be disabled
 
 ## [v1.2] - 2019-11-05
 Attention: Installing this update will create new MUMIE acounts for all users. Old MUMIE tasks will keep working, but they wont share a grade pool with newly created ones.